move FieldNames in project structure
