+++ src/main/java/com/brandonlenz/iso8583/definitions/fields/VliDefinition.java
@@ -1,6 +1,6 @@
 package com.brandonlenz.iso8583.definitions.fields;
 
-import com.brandonlenz.iso8583.definitions.names.FieldName;
+import com.brandonlenz.iso8583.definitions.fields.names.FieldName;
 import com.brandonlenz.iso8583.structure.content.ContentType;
 import com.brandonlenz.iso8583.structure.encoding.Encoding;
