+++ src/main/java/com/brandonlenz/iso8583/definitions/fields/VliFieldDefinition.java
@@ -1,7 +1,7 @@
 package com.brandonlenz.iso8583.definitions.fields;
 
 import com.brandonlenz.iso8583.building.fields.VliDataFieldBuilder;
-import com.brandonlenz.iso8583.definitions.names.FieldName;
+import com.brandonlenz.iso8583.definitions.fields.names.FieldName;
 import com.brandonlenz.iso8583.fields.VliField;
 import com.brandonlenz.iso8583.parsing.fields.VliFieldParser;
 import com.brandonlenz.iso8583.structure.content.ContentType;
