+++ src/main/java/com/brandonlenz/iso8583/definitions/messages/MessageDefinition.java
@@ -1,8 +1,7 @@
 package com.brandonlenz.iso8583.definitions.messages;
 
 import com.brandonlenz.iso8583.definitions.fields.FieldDefinition;
-import com.brandonlenz.iso8583.definitions.names.FieldName;
-import java.util.List;
+import com.brandonlenz.iso8583.definitions.fields.names.FieldName;
 import java.util.Map;
 
 public interface MessageDefinition { //TODO: rename as ProtocolDefinition? Abstract MessageDefinition seperately?
