+++ src/main/java/com/brandonlenz/iso8583/definitions/messages/SampleIso8583MessageDefinition.java
@@ -5,7 +5,7 @@ import com.brandonlenz.iso8583.definitions.fields.FieldDefinition;
 import com.brandonlenz.iso8583.definitions.fields.FixedFieldDefinition;
 import com.brandonlenz.iso8583.definitions.fields.VliDefinition;
 import com.brandonlenz.iso8583.definitions.fields.VliFieldDefinition;
-import com.brandonlenz.iso8583.definitions.names.FieldName;
+import com.brandonlenz.iso8583.definitions.fields.names.FieldName;
 import com.brandonlenz.iso8583.structure.content.ContentType;
 import com.brandonlenz.iso8583.structure.encoding.Encoding;
 import java.util.HashMap;
