+++ src/main/java/com/brandonlenz/iso8583/fields/DataField.java
@@ -1,7 +1,7 @@
 package com.brandonlenz.iso8583.fields;
 
 import com.brandonlenz.iso8583.definitions.fields.FieldDefinition;
-import com.brandonlenz.iso8583.definitions.names.FieldName;
+import com.brandonlenz.iso8583.definitions.fields.names.FieldName;
 import com.brandonlenz.iso8583.structure.format.Format;
 
 public abstract class DataField {
