+++ src/test/java/com/brandonlenz/iso8583/building/fields/VliDataFieldBuilderTest.java
@@ -2,7 +2,7 @@ package com.brandonlenz.iso8583.building.fields;
 
 import com.brandonlenz.iso8583.definitions.fields.VliDefinition;
 import com.brandonlenz.iso8583.definitions.fields.VliFieldDefinition;
-import com.brandonlenz.iso8583.definitions.names.FieldName;
+import com.brandonlenz.iso8583.definitions.fields.names.FieldName;
 import com.brandonlenz.iso8583.fields.VliField;
 import com.brandonlenz.iso8583.structure.content.ContentType;
 import com.brandonlenz.iso8583.structure.encoding.Encoding;
