+++ src/test/java/com/brandonlenz/iso8583/fields/BitmapTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.brandonlenz.iso8583.definitions.fields.BitmapDefinition;
-import com.brandonlenz.iso8583.definitions.names.FieldName;
+import com.brandonlenz.iso8583.definitions.fields.names.FieldName;
 import com.brandonlenz.iso8583.structure.encoding.Encoding;
 import java.util.Arrays;
 import org.junit.jupiter.api.Assertions;
